# Make the redis embedding dynamic
- Remove the redis.submodule directory containing the redis code from the repo.
- Update the build_scripts/update_redis_server.py script to download the redis version specified in the REDIS_VERSION environment variable to the redis.submodule directory.
- Update the screwdriver.yaml to run the build_scripts/update_redis_server.py script after the `init_os` step of all of the jobs (so the jobs have the specified version of the redis source code before the package is built or used).
- Update the setup.py to run the same logic as the build_scripts/update_redis_server.py script if the redis.submodule directory is not present.  This allows tox runs and pip installs from the repo root to continue working (which is needed for the tests to run via tox)

+++ build_scripts/update_redis_server.py
@@ -1,17 +1,20 @@
 #!/usr/bin/env python3
 """ Update the redis.submodule in the redislite repo to the latest stable version """
 import os
+import pathlib
 import shutil
 import urllib.request
 import tarfile
 import tempfile
 
 
-url = 'http://download.redis.io/releases/redis-6.2.6.tar.gz'
+redis_version = os.environ.get('REDIS_VERSION', '6.2.7')
+url = f'http://download.redis.io/releases/redis-{redis_version}.tar.gz'
 
 
 if __name__ == "__main__":
-    shutil.rmtree('redis.submodule')
+    if pathlib.Path('redis_submodule').exists():
+        shutil.rmtree('redis.submodule')
     with tempfile.TemporaryDirectory() as tempdir:
         print(f'Downloading {url} to temp directory {tempdir}')
         ftpstream = urllib.request.urlopen(url)
